requirements.txt
Flask==3.1.0
Flask_Bcrypt==1.0.1
flask_sqlalchemy==3.1.1
PyJWT==2.10.1
SQLAlchemy==2.0.36
